MQ client and server interfacing O2CDB
The user can run an MQ server/client passing CDB objects. The client
request CDB objects by specifying in the request the calibration type
("DET/Calib/Histo" in the example) followed by the run number. The
server replies with the corresponding CDB object (an instance of the
Condition class) wrapped in a TMessage.
Pre-emptyvely a local O2CDB must be filled with some objects by running
a macro. This and how to run the server and the client is detailed in
the README.
The server (ConditionsMQServer) and the client (ConditionsMQClient)
classes are adapted from the corresponding server and
client classes in FairRoot (parmq/ParameterMQServer and
examples/MQ/7-parameters/FairMQExample7Client).